edit API getAllValueOfUserInProjectByUserId
