+++ aspnet-core/src/Timesheet.Application/APIs/ProjectManagementBranchDirectors/ManageUserForBranchs/ManageUserForBranchAppService.cs
@@ -1,0 +1,126 @@
+﻿using Abp.Application.Services.Dto;
+using Abp.Authorization;
+using Abp.Linq.Extensions;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Office.Interop.Word;
+using Ncc;
+using Ncc.Authorization.Users;
+using Ncc.Entities;
+using Ncc.IoC;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Dynamic.Core;
+using System.Text;
+using System.Threading.Tasks;
+using Timesheet.APIs.ProjectManagementBranchDirectors.ManageUserForBranch.Dto;
+using Timesheet.DomainServices;
+using Timesheet.Entities;
+using Timesheet.Extension;
+using Timesheet.Paging;
+using Timesheet.Uitls;
+using Timesheet.Users.Dto;
+using static Ncc.Entities.Enum.StatusEnum;
+
+namespace Timesheet.APIs.ProjectManagementBranchDirectors.ManageUserProjectForBranchs
+{
+    public class ManageUserForBranchAppService : AppServiceBase
+    {
+        public ManageUserForBranchAppService(IWorkScope workScope) : base(workScope) { }
+
+
+        [HttpPost]
+        [AbpAuthorize]
+        [AbpAuthorize(Ncc.Authorization.PermissionNames.ProjectManagementBranchDirectors_ManageUserForBranchs_ViewAllBranchs, Ncc.Authorization.PermissionNames.ProjectManagementBranchDirectors_ManageUserForBranchs_ViewMyBranch)]
+        public async Task<PagedResultDto<UserProjectsDto>> GetAllUserPagging(GridParam input, long? positionId, long? branchId)
+        {
+            var isViewAll = await IsGrantedAsync(Ncc.Authorization.PermissionNames.ProjectManagementBranchDirectors_ManageUserForBranchs_ViewAllBranchs);
+            var currentUserBranch = WorkScope.GetAll<User>()
+                                    .Where(s => s.Id == AbpSession.UserId)
+                                    .Select(s => s.BranchId).FirstOrDefault();
+
+            var qprojectUsers = from pu in WorkScope.GetAll<ProjectUser>().Where(s => s.User.IsActive == true)
+                                join p in WorkScope.GetAll<Project>().Where(s => s.Status == ProjectStatus.Active) on pu.ProjectId equals p.Id
+                                where pu.Type != ProjectUserType.DeActive
+                                select new
+                                {
+                                    pu.ProjectId,
+                                    p.Code,
+                                    p.Name,
+                                    pu.UserId,
+                                    pu.Type
+                                };
+
+            var query = from u in WorkScope.GetAll<User>()
+                        join pu in qprojectUsers on u.Id equals pu.UserId into pusers
+                        join mu in WorkScope.GetAll<User>() on u.ManagerId equals mu.Id into muu
+                        select new UserProjectsDto
+                        {
+                            Id = u.Id,
+                            UserName = u.UserName,
+                            Name = u.Name,
+                            Surname = u.Surname,
+                            FullName = u.FullName,
+                            Address = u.Address,
+                            IsActive = u.IsActive,
+                            EmailAddress = u.EmailAddress,
+                            PhoneNumber = u.PhoneNumber,
+                            ProjectUsers = pusers.Select(s => new PUDto
+                            {
+                                ProjectId = s.ProjectId,
+                                ProjectName = s.Name,
+                                ProjectCode = s.Code,
+                                ProjectUserType = s.Type
+                            }).ToList(),
+                            Type = u.Type,
+                            Level = u.Level,
+                            UserCode = u.UserCode,
+                            BranchDisplayName = u.Branch.DisplayName,
+                            BranchId = u.BranchId,
+                            AvatarPath = u.AvatarPath,
+                            ManagerId = u.ManagerId,
+                            Sex = u.Sex,
+                            CreationTime = u.CreationTime,
+                            ManagerName = muu.FirstOrDefault() != null ? muu.FirstOrDefault().FullName : "",
+                            ManagerAvatarPath = muu.FirstOrDefault() != null ? muu.FirstOrDefault().AvatarPath : "",
+                            PositionId = u.Position.Id,
+                            PositionName = u.Position.Name,
+                            ProjectCount = pusers.Count(),
+                        };
+            if (!isViewAll)
+            {
+                query = query.WhereIf(positionId != null, s => s.PositionId == positionId)
+                             .Where(s => s.BranchId == currentUserBranch)
+                             .OrderByDescending(s => s.CreationTime);
+            }
+            else
+            {
+                query = query.WhereIf(positionId != null, s => s.PositionId == positionId)
+                             .WhereIf(branchId != null, s => s.BranchId == branchId)
+                             .OrderByDescending(s => s.CreationTime);
+            }
+            var temp = await query.GetGridResult(query, input);
+
+            var projectIds = new HashSet<long>();
+            foreach (var user in temp.Items)
+            {
+                projectIds.UnionWith(user.ProjectUsers.Select(s => s.ProjectId));
+            }
+
+            var projects = (WorkScope.GetAll<ProjectUser>()
+                    .Where(s => projectIds.Contains(s.ProjectId) && s.Type == ProjectUserType.PM)
+                    .Select(s => new { s.ProjectId, s.User.FullName })
+                    .GroupBy(s => s.ProjectId))
+                    .Select(s => new { s.Key, pms = s.Select(f => f.FullName).ToList() }).ToList();
+
+            foreach (var user in temp.Items)
+            {
+                foreach (var pu in user.ProjectUsers)
+                {
+                    pu.Pms = projects.Where(s => s.Key == pu.ProjectId).Select(s => s.pms).FirstOrDefault();
+                }
+            }
+            return new PagedResultDto<UserProjectsDto>(temp.TotalCount, temp.Items);
+        }
+    }
+}
