+++ aspnet-core/src/Timesheet.Core/Authorization/PermissionNames.cs
@@ -332,11 +332,10 @@ namespace Ncc.Authorization
         public const string TeamBuilding_Project_SelectProjectTeamBuilding = "TeamBuilding.Project.SelectProjectTeamBuilding";
 
         public const string ProjectManagementBranchDirectors = "ProjectManagementBranchDirectors";
+        public const string ProjectManagementBranchDirectors_ManageUserForBranchs = "ProjectManagementBranchDirectors.ManageUserForBranchs";
+        public const string ProjectManagementBranchDirectors_ManageUserForBranchs_ViewAllBranchs = "ProjectManagementBranchDirectors.ManageUserForBranchs.ViewAllBranchs";
+        public const string ProjectManagementBranchDirectors_ManageUserForBranchs_ViewMyBranch = "ProjectManagementBranchDirectors.ManageUserForBranchs.ViewMyBranch";
         public const string ProjectManagementBranchDirectors_ManageUserProjectForBranchs = "ProjectManagementBranchDirectors.ManageUserProjectForBranchs";
-        public const string ProjectManagementBranchDirectors_ManageUserProjectForBranchs_ViewAllBranchs = "ProjectManagementBranchDirectors.ManageUserProjectForBranchs.ViewAllBranchs";
-        public const string ProjectManagementBranchDirectors_ManageUserProjectForBranchs_ViewMyBranch = "ProjectManagementBranchDirectors.ManageUserProjectForBranchs.ViewMyBranch";
-
-
     }
 
     public class GrantPermissionRoles
@@ -649,9 +648,10 @@ namespace Ncc.Authorization
                     PermissionNames.TeamBuilding_Project,
                     PermissionNames.TeamBuilding_Project_SelectProjectTeamBuilding,
                     PermissionNames.ProjectManagementBranchDirectors,
+                    PermissionNames.ProjectManagementBranchDirectors_ManageUserForBranchs,
+                    PermissionNames.ProjectManagementBranchDirectors_ManageUserForBranchs_ViewAllBranchs,
+                    PermissionNames.ProjectManagementBranchDirectors_ManageUserForBranchs_ViewMyBranch,
                     PermissionNames.ProjectManagementBranchDirectors_ManageUserProjectForBranchs,
-                    PermissionNames.ProjectManagementBranchDirectors_ManageUserProjectForBranchs_ViewAllBranchs,
-                    PermissionNames.ProjectManagementBranchDirectors_ManageUserProjectForBranchs_ViewMyBranch,
     }
             },
 
@@ -803,8 +803,9 @@ namespace Ncc.Authorization
                     PermissionNames.ManageWorkingTime_ViewDetail,
                     PermissionNames.ManageWorkingTime_Approval,
                     PermissionNames.ProjectManagementBranchDirectors,
+                    PermissionNames.ProjectManagementBranchDirectors_ManageUserForBranchs,
+                    PermissionNames.ProjectManagementBranchDirectors_ManageUserForBranchs_ViewMyBranch,
                     PermissionNames.ProjectManagementBranchDirectors_ManageUserProjectForBranchs,
-                    PermissionNames.ProjectManagementBranchDirectors_ManageUserProjectForBranchs_ViewMyBranch,
                 }
             },
         };
@@ -1125,9 +1126,10 @@ namespace Ncc.Authorization
             new SystemPermission{ Name =  PermissionNames.TeamBuilding_Project_SelectProjectTeamBuilding, MultiTenancySides = MultiTenancySides.Host, DisplayName = "Select project team building"},
 
             new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors, MultiTenancySides = MultiTenancySides.Host, DisplayName = "Project Management For Branch Directors"},
+            new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors_ManageUserForBranchs, MultiTenancySides = MultiTenancySides.Host, DisplayName = "Manage User For Branchs"},
+            new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors_ManageUserForBranchs_ViewAllBranchs, MultiTenancySides = MultiTenancySides.Host, DisplayName = "View User from all branchs"},
+            new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors_ManageUserForBranchs_ViewMyBranch, MultiTenancySides = MultiTenancySides.Host, DisplayName = "View User from my branch"},
             new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors_ManageUserProjectForBranchs, MultiTenancySides = MultiTenancySides.Host, DisplayName = "Manage User Project For Branchs"},
-            new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors_ManageUserProjectForBranchs_ViewAllBranchs, MultiTenancySides = MultiTenancySides.Host, DisplayName = "View User Project from all branchs"},
-            new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors_ManageUserProjectForBranchs_ViewMyBranch, MultiTenancySides = MultiTenancySides.Host, DisplayName = "View User Project from my branch"},
 
         };
 
@@ -1730,13 +1732,14 @@ namespace Ncc.Authorization
              },new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors, MultiTenancySides = MultiTenancySides.Host, DisplayName = "Project Management For Branch Directors",
                  Childrens = new List<SystemPermission>()
                  {
-                     new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors_ManageUserProjectForBranchs, MultiTenancySides = MultiTenancySides.Host, DisplayName = "Manage User Project For Branchs",
+                     new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors_ManageUserForBranchs, MultiTenancySides = MultiTenancySides.Host, DisplayName = "Manage User For Branchs",
                          Childrens = new List<SystemPermission>()
                         {
-                         new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors_ManageUserProjectForBranchs_ViewAllBranchs, MultiTenancySides = MultiTenancySides.Host, DisplayName = "View User Project from all branchs"},
-                         new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors_ManageUserProjectForBranchs_ViewMyBranch, MultiTenancySides = MultiTenancySides.Host, DisplayName = "View User Project from my branch"},
+                         new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors_ManageUserForBranchs_ViewAllBranchs, MultiTenancySides = MultiTenancySides.Host, DisplayName = "View User from all branchs"},
+                         new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors_ManageUserForBranchs_ViewMyBranch, MultiTenancySides = MultiTenancySides.Host, DisplayName = "View User from my branch"},
                         }
-                     }
+                     },
+                     new SystemPermission{ Name =  PermissionNames.ProjectManagementBranchDirectors_ManageUserProjectForBranchs, MultiTenancySides = MultiTenancySides.Host, DisplayName = "Manage User Project For Branchs" },
                  }
              }
         };
